🚚 Moved tests to components folder
